Servidor Node.Js agregeado al proyecto, agregada respues de error404, index con pug y organizado segun modelo vista controlador
