+++ .gitignore.txt
@@ -1,0 +1,2 @@
+# ignorando la carpeta de node_modules
+node_modules/
