--- GestorBirthday/gestor.css
@@ -1,123 +1,0 @@
-/*Ventana emergente*/
-#model{
-    display: none;
-    background-color: rgba(0, 0, 0, 0.4);
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    align-items: center;/*asigna en el centro si*/
-    justify-content: center; /*el display esta en fixed*/
-}
-#contenido_model{
-    position: relative;  /*para el boton de cancelar*/
-    border: solid white;
-    background-color:rgb(51, 100, 207);
-    width: 290px;
-    margin: 15% auto;
-    padding: 0px 20px 20px 20px;
-    border-radius: 20px;
-}
-
-/*icono de cerrar*/
-#cerrar{    
-    display: inline;
-    float: right;
-    color: white;
-    font-size: 28px;
-    font-weight: bold;
-    /*margin-left: 329px;*/
-}
-#cerrar:hover{
-    color: red;
-    cursor: pointer;
-}
-
-/*Botones*/
-.form{
-    background-color: rgb(51, 100, 207);
-    color: white;
-    font-size: 130%;
-    border-radius: 10px;
-    border: none;
-    padding-top: 4px;
-    padding-bottom: 4px;
-}
-#cancelar{
-    position: absolute;
-    right: 35px;
-    bottom: 19px;
-}
-.operacion{
-    background-color: rgb(51, 100, 207);
-    color: white;
-    font-size: 130%;
-    border-radius: 20px;
-    border: none;
-    padding-top: 4px;
-    padding-bottom: 4px;
-    margin: 15px;
-    width: auto;
-}
-.operacion:hover, .form:hover{
-    background-color: rgb(48, 127, 237);    
-}
-.apagado{
-    background-color: rgb(58, 78, 122);
-    color: rgb(138, 136, 136);
-    font-size: 130%;
-    border-radius: 20px;
-    border: none;
-    padding-top: 4px;
-    padding-bottom: 4px;
-    margin: 15px;
-    width: auto;
-}
-.apagado:hover{
-    cursor: default;
-}
-
-/*Inputs*/
-input[type = 'text'], input[type = 'date'] {
-    width: 175px;
-    margin-left: 30px;
-}
-input[type = 'date']{
-    margin-left: 43.8px;
-    width: 179px;
-}
-
-/*tabla*/
-th,td {    
-    text-align: center;
-    display: inline-block;
-    border: white solid;
-    color: white;
-    width: 40%; /*200px*/ 
-    cursor: pointer;
-}
-th{
-    cursor:default;
-    background-color: rgb(48, 127, 237);
-}
-table{
-    width: 100%; 
-    border-collapse: collapse;
-}
-/*extras*/
-#cuerpo{
-    text-align: center;
-}
-#persona{
-    color: white;
-}
-#asterisco{
-    color: red;
-}
-img{
-    position: absolute;
-    width: 24px;
-    margin-left: 5px;
-    cursor: pointer;
-}
