--- README.md
@@ -1,4 +1,0 @@
-# Cronometro-en-JS
-Un cronometro creado en JS o para la web
-
-Es un proyecto bastante sencillo, mas bien para pulir habilidades en lo que respecta a JS, CSS, y HTML asi como aplicar nuevas tecnicas y quizas en un futuro Node.js para entrenar. Se basa en un connocrometro estandar, se detiene al precionar un boton, y continua al precionar el mismo, aunque se ha ido incrementando las opciones hasta agregar un gestor de fechas de cumpleaños y un temporizador
