--- index.css
@@ -1,44 +1,0 @@
-/*cronometro y almacen de marcas*/
-#crono{
-    padding-top: 25px;
-    font-size: 400%;
-}
-#almacen{
-    position: fixed;
-    background-color: rgb(29, 29, 177);
-    border-bottom: 2px solid rgb(30, 30, 30);
-    border-top: 2px solid rgb(30, 30, 30);
-    height: 40%;
-    right: 0;
-    width: 100%;
-    top: 210px;
-    overflow-y: scroll;
-}
-
-/*Botones de opciones*/
-#cajaBotones{
-    position: fixed;
-    width: 97%;
-    bottom: 10%;
-    right: 1%;
-}
-div button{
-    background-color: rgb(29, 29, 177);
-    color: white;
-    font-size: 200%;
-    border-radius: 15px;
-    width: auto; /*32.75%*/
-    border: 1px solid white;
-    transition: background-color 0.3s;
-}
-div button:hover{
-    background-color: rgb(51, 87, 207);
-}
-
-/*Extras*/
-p{
-    font-size: 150%;
-}
-div{
-    text-align: center;
-}
