--- index.html
@@ -1,53 +1,0 @@
-<!DOCTYPE html>
-<html lang = "es">   
-    <!--
-        TODO:
-        1: Cambiar esta pestaña por otra aparte con la funcionalidad de cronometro
-        y establecer una pagina central para acceder a los 3 servicios
-        2: Agregar un carrucel al index con los 3 servicios
-     -->
-    <head>
-        <title>Index de Cronometro</title>
-        <meta charset="UTF-8">
-
-        <link rel="stylesheet" href="index.css">
-        <link rel="stylesheet" href="Predeterminado.css">
-        
-    </head>
-
-    <body>
-        <header>
-            <nav class="navegacion">
-                <a href="Temporizador/Temporizador.html"><button class="linkButton">Temporizador</button></a>
-                <a href="GestorBirthday/gestor.html"><button class="linkButton">Gestor de cumpleaños</button></a>                
-            </nav>
-        </header>
-
-        <main>
-            
-            <div>
-                <p id = 'crono'>00:00</p>                                 
-            </div> 
-            
-            <div id = 'cajaBotones'>
-                <button id = "Resetear">Reiniciar</button>                   
-                <button id = "Start">Iniciar</button>                   
-                <button id = "Marcar">Marcar</button> 
-            </div>
-        </main>        
-        <div id = "almacen"></div>
-        <footer>
-            <ul>
-                <li id = 'linkdin'><a target = '_blank' href="https://www.linkedin.com/in/derik-gonz%C3%A1lez-mart%C3%ADn-6b73242a3/">
-                Linkdin</a></li>
-                <li id = "github"><a target = '_blank' href="https://github.com/derikgm">Github</a></li>
-            </ul> 
-        </footer> 
-        
-    </body>
-
-<!--Seccion JS-->
-<script src="cronometro.js"></script>
-<script src="hora.js"></script>
-</html>
-
