+++ models/db-config.json
@@ -1,0 +1,9 @@
+{
+    "mysql" : {
+        "host": "localhost",
+        "port": 5432,
+        "user": "",
+        "pass": "",
+        "database": ""
+    }
+}
