+++ public/Error404.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Error 404</title>
+    <link rel="stylesheet" href="/css/default//Predeterminado.css">
+    <link rel="stylesheet" href="/css/default/error404.css">
+</head>
+<body>
+    <main>
+        <div>
+            <h1 style="color: rgb(231, 71, 71);">Error 404: file not found!</h1>
+            <h2>Opps! Parece que la seccion que buscas no se encuentra</h2>  
+            <a id = 'back' href="/">volver</a>          
+        </div>
+        
+    </main>
+</body>
+</html>
