+++ public/Temporizador.html
@@ -1,0 +1,53 @@
+<!DOCTYPE html><!--Completa-->
+<html lang = "es">
+    <head>
+        <meta charset="UTF-8">
+        <title>Document</title>
+
+        <link rel="stylesheet" href="/css/default/Predeterminado.css">
+        <link rel="stylesheet" href="/css/temporizador/Temporizador.css">
+
+    </head>
+    <body>
+        <header>
+            <nav class = "navegacion">
+                <a href="/"><button id="home">Inicio</button></a>
+                <a href="/Cronometro.html"><button>Cronometro</button></a>
+                <a href="/Gestor.html"><button>Gestor de cumpleaños</button></a>                
+            </nav>
+        </header>
+    <main>
+        <br><br>
+        <div id = "caja"></div>
+            <p id = "temp">0:00:00</p>
+        <div id = 'cuerpo'>
+            <form id = "formulario"> 
+                <label for="h">H:</label>           
+                <input type="number" id="h" name="hora" value = "0" min='0' max="30" placeholder="hora">
+                <label for="m" id="min">M:</label>
+                <input type="number" id="m" name="minuto" value = "0" min='0' max="59" placeholder="min">
+                <label for="s">S:</label>
+                <input type="number" id="s" name = "segundo" value = "0" min='0'max="59" placeholder="seg"><br><br>
+                <input type="submit" class = 'operar' value="Iniciar" id="submit"> 
+            </form>
+            <button class = 'operar' id = "reiniciar_boton">Reiniciar</button>    
+            <button class = 'operar' id = "parar_boton">Parar alarma</button>
+        </div>        
+    </main>
+     <footer> 
+            <ul>         
+                <li id = 'linkdin'><a target = '_blank' href="https://www.linkedin.com/in/derik-gonz%C3%A1lez-mart%C3%ADn-6b73242a3/">
+                Linkdin</a></li>
+                <li id = "github"><a target = '_blank' href="https://github.com/derikgm">Github</a></li>
+            </ul>  
+        </footer>
+        <audio id = "alarma" >
+        <source src="/audio/Alarm-Clock-Short-chosic.com_.mp3" type="audio/mp3">
+        </audio>
+    </body>
+
+    <!--Seccion JS-->
+    <script src="/js/temporizador/Temporizador.js"></script>
+    <script src="/js/default/hora.js"></script>
+
+</html>
