+++ public/css/default/Predeterminado.css
@@ -1,0 +1,96 @@
+/*Seccion del cuerpo*/
+body{
+    background-color: rgb(51, 51, 207);
+}
+footer{
+    text-align: left;
+}
+
+/*Seccion de links*/
+a{
+    color: white;
+    text-decoration: none;
+}
+li a:hover{
+    color: aqua;
+    text-decoration: underline;
+}
+a:active{
+    color: blueviolet;
+}
+a:link{
+    color: aqua;
+}
+#linkdin{
+    list-style-image: url('/images/linkedin-logo.png');
+}
+#github{
+    list-style-image: url('/images/github logo.png');
+}
+footer{
+    background-color: darkblue;
+    position: fixed; /*Posicion fija*/
+    width: 100%;
+    height: 8%;    /*NO olvidar que se debe de establecer las dimenciones*/
+    bottom: 0;
+    left: 0;
+}
+
+/*Barra de navegacion superior*/
+.navegacion{
+    background-color: rgb(59, 59, 233); 
+    text-align: center;
+    border-bottom: white solid;
+    position: fixed;
+    width: 100%;
+    top: 0;
+    left: 0;
+}
+.navegacion a {
+    background-color: rgb(59, 59, 233); 
+    text-align: center;
+}
+a button{
+    border: none;
+    background-color: rgb(59, 59, 233);
+    color: white;
+    width: 45%;
+    font-size: 120%;
+    transition: 0.35s;
+}
+a button:hover{
+    background-color:rgb(59, 100, 233);
+}
+#home{
+    width: auto;
+}
+
+/*Elementos estandars*/
+p{
+    color: white;
+}
+li{
+    font-size: 115%;
+    margin-right: 50px;    
+}
+ul{
+    display: flex;    
+    flex-direction: row;
+}
+label{
+    color: white;
+}
+#reloj{
+    font-size: large;
+    position: absolute;
+    display: inline;   /*Transforma bloques de objetos en lineas*/ 
+    top: 3%;
+    right: 10px;
+}
+button, input[type = 'submit']{
+    cursor: pointer;
+}
+h1{
+    color: white;
+    text-align: center;
+}
