+++ views/index.pug
@@ -1,0 +1,31 @@
+doctype html
+html(lang="en")
+    head
+        meta(charset="UTF-8")
+        meta(name="viewport", content="width=device-width, initial-scale=1.0")
+        title index y portada
+        link(rel="stylesheet", href="/css/default/Predeterminado.css")
+        link(rel="stylesheet", href="/css/default/index.css")
+    body
+        main 
+            h1 Gestor de tiempo
+
+            div
+                picture 
+                    a(href="/Cronometro.html") 
+                        img(src="/images/Cronometro.jfif", width = 250,alt="Cronometro, enlace a la pagina de cronometro")
+                picture 
+                    a(href="/Temporizador.html") 
+                        img(src="/images/Temporizador.jfif", width = 250,alt="Temporizador, enlace a la pagina de temporizador")
+                picture 
+                    a(href="/Gestor.html") 
+                        img(src="/images/birthday date.jfif", width = 250,alt="Gestor, enlace a la pagina de gestor")
+        footer
+            ul
+                li#linkdin
+                    a(target='_blank', href="https://www.linkedin.com/in/derik-gonz%C3%A1lez-mart%C3%ADn-6b73242a3/") Linkdin
+                li#github
+                    a(target='_blank', href="https://github.com/derikgm") Github
+             
+
+script(src="/js/default/hora.js")
